Added German translation
German translation for the interface.
I was looking for long time for some way to integrate La Nordica stove. The german community is very active and I hope we can give your project a boost. I hope we can some day tap into the BUS of the stove to eliminate the need of connecting it to the cloud.