Add check for optionnal parameters
Some keys in register are optionals (key is present in software but option is not physically present on stove).
Added check of "xxxx_enable" for check availability of register.
As not all status are presents on my Nobis stove, i've only set an "enable_key" for know status (when an available key is present).